Formatter: port the old formatter's test suite to the query formatter
The old imperative formatter's 52 end-to-end tests move from fmt.rs to
the new fmt/tests.rs, now running against format_document_query. The
expected outputs are re-blessed where the two formatters deliberately
disagree (input-driven layout instead of forced breaking, no line-width
breaking, no trailing-comma management, tight empty braces, dropped
leading blank lines, guaranteed trailing newline), so the suite records
the before/after difference. The query formatter's own overlapping
end-to-end tests in rules.rs are removed; only its two leaf-rule tests
survive, appended to the ported suite, because no old test covers those
constructs.

Two rule fixes surfaced by the port: a dot no longer glues onto a bare
integer literal (42.log(x) re-lexes as `42.` `log` and would change
the expression), and stray error-recovery token children of a document
or braced body keep their input line structure instead of being broken
one token per line.
